// src/camps.generated.ts
// Generated by npm run sync:camps. Do not edit manually.
export const camps = [
  {
    "id": "CAMP-07",
    "title": "程诗楠の知识",
    "href": "https://wanhuxian.feishu.cn/wiki/Lr9TwXZToi3oVBkDDAmcigGQnuh",
    "nodeToken": "Lr9TwXZToi3oVBkDDAmcigGQnuh",
    "documentCount": 0
  },
  {
    "id": "CAMP-15",
    "title": "罗吉煜",
    "href": "https://wanhuxian.feishu.cn/wiki/FO60wRo2FieM9CkOS5RcR7SEnBe",
    "nodeToken": "FO60wRo2FieM9CkOS5RcR7SEnBe",
    "documentCount": 1
  },
  {
    "id": "CAMP-05",
    "title": "张佳宁的视角",
    "href": "https://wanhuxian.feishu.cn/wiki/J6AUwNorFidCkakck2vcKHCenTb",
    "nodeToken": "J6AUwNorFidCkakck2vcKHCenTb",
    "documentCount": 2
  },
  {
    "id": "CAMP-11",
    "title": "张娇",
    "href": "https://wanhuxian.feishu.cn/wiki/ILWywqqslig2pFkqdKAcI1d6nmh",
    "nodeToken": "ILWywqqslig2pFkqdKAcI1d6nmh",
    "documentCount": 2
  },
  {
    "id": "CAMP-04",
    "title": "俱渊博的分享圈（更新ing）",
    "href": "https://wanhuxian.feishu.cn/wiki/YQDRwdks8iUUM1ksFN0ciCdNn6b",
    "nodeToken": "YQDRwdks8iUUM1ksFN0ciCdNn6b",
    "documentCount": 3
  },
  {
    "id": "CAMP-02",
    "title": "严茜个人分享（GEO相关）",
    "href": "https://wanhuxian.feishu.cn/wiki/HohMwnGlXiJyFUk280YcwXWunQe",
    "nodeToken": "HohMwnGlXiJyFUk280YcwXWunQe",
    "documentCount": 3
  },
  {
    "id": "CAMP-12",
    "title": "贺李蒙的思考",
    "href": "https://wanhuxian.feishu.cn/wiki/AEs7wZVXpi9pXFkHxlIc1siZnec",
    "nodeToken": "AEs7wZVXpi9pXFkHxlIc1siZnec",
    "documentCount": 5
  },
  {
    "id": "CAMP-08",
    "title": "苏苏的小思考",
    "href": "https://wanhuxian.feishu.cn/wiki/Z1EHwOgEzibU8ekkvdycI91Xnfu",
    "nodeToken": "Z1EHwOgEzibU8ekkvdycI91Xnfu",
    "documentCount": 5
  },
  {
    "id": "CAMP-13",
    "title": "Livia",
    "href": "https://wanhuxian.feishu.cn/wiki/TkiSw3YFsi0ee3kF8x4c5TAtnYg",
    "nodeToken": "TkiSw3YFsi0ee3kF8x4c5TAtnYg",
    "documentCount": 8
  },
  {
    "id": "CAMP-01",
    "title": "小陈的GEO心流时刻",
    "href": "https://wanhuxian.feishu.cn/wiki/RIZTw6ordiWUW2kUwbGcit6knmb",
    "nodeToken": "RIZTw6ordiWUW2kUwbGcit6knmb",
    "documentCount": 11
  },
  {
    "id": "CAMP-09",
    "title": "姚金刚",
    "href": "https://wanhuxian.feishu.cn/wiki/XQgAwxpP6iKqgGkmpIfcc28Enee",
    "nodeToken": "XQgAwxpP6iKqgGkmpIfcc28Enee",
    "documentCount": 13
  },
  {
    "id": "CAMP-06",
    "title": "宋斌斌的进步之路",
    "href": "https://wanhuxian.feishu.cn/wiki/Os9aw9VYbiZgl9kB0nBceB8snwh",
    "nodeToken": "Os9aw9VYbiZgl9kB0nBceB8snwh",
    "documentCount": 15
  },
  {
    "id": "CAMP-03",
    "title": "刘程路@移山科技",
    "href": "https://wanhuxian.feishu.cn/wiki/WQi1wQjOGijo2gk5i2rcpdZsnAd",
    "nodeToken": "WQi1wQjOGijo2gk5i2rcpdZsnAd",
    "documentCount": 28
  },
  {
    "id": "CAMP-10",
    "title": "Asuka",
    "href": "https://wanhuxian.feishu.cn/wiki/WDt0wQA8BivozPk7Z9GcrmUVnyg",
    "nodeToken": "WDt0wQA8BivozPk7Z9GcrmUVnyg",
    "documentCount": 45
  },
  {
    "id": "CAMP-14",
    "title": "潇斌的SPACE",
    "href": "https://wanhuxian.feishu.cn/wiki/Xg2Mw21B9iWzDkkL8NscouZingb",
    "nodeToken": "Xg2Mw21B9iWzDkkL8NscouZingb",
    "documentCount": 56
  },
  {
    "id": "CAMP-16",
    "title": "Sora",
    "href": "https://wanhuxian.feishu.cn/wiki/RXduwIcjFi9zcbkBsREcM1g5n1g",
    "nodeToken": "RXduwIcjFi9zcbkBsREcM1g5n1g",
    "documentCount": 56
  }
] as const